Structure update, level pipeline outlined
Renamed Map to Screen, added TileMap which handles sets of tiles for
later rendering.

Added Level for handling sets of TileMaps. A level has a marker that
stores which TileMap is currently active which will get rendered.

Changed a lot of the core component variables to pointers for easier
passing around to other classes.

Overall, nothing is different when testing but the code is vastly
simpler and clearly structured.