Added EntityManager drawing and fixed shooting
EntityManager now handles the drawing of entities. It also has a very
crude method for rendering each type of entity in the right order (4
loops which each pick out only a specific type to render. Will be
replaced with sorting in the future).

Moved time to live checking to EntityManager in order to ensure they are
deleted properly from the entity manager list.

Improved the naming convention in Sprite.

Removed memes.